refactor dataset wrapper into basic and default
